Move everything around for the quiz.app naming
