+++ .github/workflows/ploomber-cloud.yaml
@@ -25,7 +25,6 @@ jobs:
       - name: Deploy
         env:
           PLOOMBER_CLOUD_KEY: ${{ secrets.PLOOMBER_CLOUD_KEY }}
-          FLASK_APP: quiz/app.py
         run: |
           ploomber-cloud deploy --watch-incremental
         # The --watch-incremental flag will print deployment status and deployment
