+++ quiz/app.py
@@ -5,19 +5,19 @@ Copyright © 2024 J. Michael McGarrah <mcgarrah@gmail.com>
 
 from flask import Flask
 from flask_migrate import Migrate
-from quiz import __version__, __build_date__
+from . import __version__, __build_date__
 
 # Import models and routes using absolute imports
-from modules.models import db, Category, Question
-from modules.routes_main import ( import_questions, export_questions, clear_questions,
-                                    add_question, delete_question,
-                                    home, edit_questions,
-                                    edit_categories, add_category, delete_category )
-from modules.routes_quiz import quiz, check_answers
-from modules.routes_settings import settings, update_settings
+from quiz.modules.models import db, Category, Question
+from quiz.modules.routes_main import ( import_questions, export_questions, clear_questions,
+                                        add_question, delete_question,
+                                        home, edit_questions,
+                                        edit_categories, add_category, delete_category )
+from quiz.modules.routes_quiz import quiz, check_answers
+from quiz.modules.routes_settings import settings, update_settings
 
 # Initialize the Flask application
-app = Flask(__name__, template_folder='../templates')
+app = Flask(__name__)
 
 @app.context_processor
 def inject_version():
