+++ quiz/templates/footer.html
