+++ quiz/templates/settings.html
